Add files via upload
Relative Orientation: Let's assume the orientation factor affects the ability of one molecule to transfer energy to another. If they are aligned well, energy transfer is more efficient. We can represent this as a random factor that reduces the energy transferred between molecules.

Environment: We could simulate a basic environmental factor, like temperature. If the system is hotter, the molecules might be more excited and thus able to accept more energy.