Add files via upload
Added a spectral_absorption_coefficients parameter to the Molecule constructor. This parameter is a dictionary that associates different wavelengths of light with absorption coefficients. When light is absorbed, the energy of the light is multiplied by the coefficient for its wavelength.

Added a wavelength parameter to the absorb_light and pulse_light methods. 

Added quantum mechanical behavior to the absorb_light method. After the absorbed energy is calculated